Add images:lqip command
This requires Imagemagick to be installed on the server, but then so
does `image:color` for palette generation. However, this command uses
`exec()`, which might be disabled on the server. If so, we can switch
to the Imagick library; there might be some performance degradation,
but it shouldn't be a big deal.