Refactor python:export
Previously, this command was meant more for transferring data between
servers than for facilitating communication between PHP and Python
components of this data service. We created a good dump infrastructure
in the data-aggregator, which can make the old code obsolete. Once the
need arises, we can transfer it here via the foundation.