Fix `insertUpdate` in MySQL builder
This is neat. We can now for sure do `findOrNew` on a database level,
which allows for bulking and is vastly faster than doing so through
Eloquent. Perfect for importing stuff from upstream APIs.

Previously, we used `insertIgnore` for safety: running `api:import`
on an existing database would not wipe computed columns. We could
have used `replace`, since it worked fine, but doing so would wipe
out local data. In effect, we didn't have a way to update the local
database with fresh upstream data.

But now, we can use the `api:import` command safely with an existing
database: if the primary key conflicts, it'll just update the row!